+++ ecs.php
@@ -12,6 +12,6 @@ return static function(ECSConfig $ecsConfig): void {
     ]);
 
     $ecsConfig->sets([
-        SetList::CRAFT_CMS_4,
+        SetList::CRAFT_CMS_5,
     ]);
 };